# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "description": "新建研究工作区的最小输入，不允许客户端指定所有者或状态。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name"],
        "title": "CreateWorkspaceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateWorkspaceRequest": {
        "description": "修改活动工作区的显示信息，研究问题将在意图确认功能中单独处理。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateWorkspaceRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStageDisplay": {
        "description": "工作区研究阶段的中文展示元数据，不作为客户端提交输入。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": ["label", "description"],
        "title": "WorkflowStageDisplay",
        "type": "object"
      },
      "WorkspaceListResponse": {
        "description": "工作区切换器使用的游标分页结果。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "WorkspaceListResponse",
        "type": "object"
      },
      "WorkspaceResponse": {
        "description": "工作区列表和详情共用的可展示字段。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "research_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Question"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_stage": {
            "$ref": "#/components/schemas/WorkspaceWorkflowStage"
          },
          "workflow_stage_display": {
            "$ref": "#/components/schemas/WorkflowStageDisplay"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "research_question",
          "status",
          "workflow_stage",
          "workflow_stage_display",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceWorkflowStage": {
        "description": "工作区在研究主流程中的阶段；英文值会写入数据库和 API。",
        "enum": [
          "draft",
          "analyzing",
          "plan_review",
          "retrieving",
          "screening",
          "collection_building",
          "researching",
          "failed"
        ],
        "title": "WorkspaceWorkflowStage",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "面向学术文献检索与 RAG 研究工作区的后端服务。",
    "title": "Academic Search API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "description": "供工作区切换器按需加载，默认只显示活动工作区。",
        "operationId": "list_workspaces_api_v1_collections_get",
        "parameters": [
          {
            "description": "是否包含已归档工作区",
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "description": "是否包含已归档工作区",
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "description": "按工作区名称或当前研究阶段搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按工作区名称或当前研究阶段搜索",
              "title": "Q"
            }
          },
          {
            "description": "上一页返回的不透明分页游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "上一页返回的不透明分页游标",
              "title": "Cursor"
            }
          },
          {
            "description": "本次最多返回的工作区数量",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "本次最多返回的工作区数量",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "搜索我的研究工作区",
        "tags": ["研究工作区"]
      },
      "post": {
        "description": "创建一个空工作区；研究问题将在进入工作区后另行确认。",
        "operationId": "create_workspace_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "创建研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
    "/api/v1/collections/{collection_id}": {
      "get": {
        "description": "读取当前用户拥有的单个活动或归档工作区。",
        "operationId": "get_workspace_api_v1_collections__collection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取研究工作区详情",
        "tags": ["研究工作区"]
      },
      "patch": {
        "description": "仅更新名称和说明，归档工作区必须先恢复才能编辑。",
        "operationId": "update_workspace_api_v1_collections__collection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "修改活动工作区信息",
        "tags": ["研究工作区"]
      }
    },
    "/api/v1/collections/{collection_id}/archive": {
      "post": {
        "description": "归档后停止新的文献写入与研究活动，已保存数据保持可恢复。",
        "operationId": "archive_workspace_api_v1_collections__collection_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "归档研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
    "/api/v1/collections/{collection_id}/restore": {
      "post": {
        "description": "恢复后工作区可继续接收后续检索和文献入库任务。",
        "operationId": "restore_workspace_api_v1_collections__collection_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "恢复研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
  }
}